backend/reports-service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
clickhouse-driver==0.2.6
minio==7.1.16
PyJWT==2.8.0
python-dotenv==1.0.0
python-dateutil==2.8.2 